refactor: package poster results, split solvers into classes, expand batch algorithms
- Move poster_results_* into src/commands/poster_results/ package
- Split poster_results_solvers.py → solver classes (BaseSolver, RawSASolver, etc.)
- Expand POSTER_BATCH_ALGORITHMS to include all MR2S variants and DnC strategies
- Add directed_edges field to result dataclasses for orientation tracking
- _run_poster_algorithm dispatches to solver classes via ALGORITHM_SOLVER_MAP
- Backward-compat via package __init__.py re-exports, old stubs removed
