+++ src/commands/poster_results/__init__.py
@@ -1,0 +1,87 @@
+from src.commands.poster_results.models import (
+    BasicResultSeries,
+    BasicSolverResult,
+    GlobalResultSeries,
+    GlobalSolverResult,
+    Mr2sResultSeries,
+    Mr2sSolverResult,
+    Mr2sTrialResult,
+    PosterResultsSummary,
+    PosterTrialResult,
+    RandomBaselineResult,
+    TimingResultSeries,
+    TrialPayload,
+    TrialTimings,
+    coerce_trial_payload,
+    payload_to_dict,
+)
+from src.commands.poster_results.cache import (
+    CachedTrialTask,
+    CachedTrialWorker,
+    cached_trial_result,
+)
+from src.commands.poster_results.partition_strategy import (
+    DncPartitionStrategy,
+    TimedPartitionStrategy,
+    _can_recurse_partition,
+    _divide_graph_with_diagnostics,
+    _embedding_estimate_with_status,
+    _estimate_physical_qubits,
+    _estimate_physical_qubits_with_status,
+    _find_partition_by_target_k_with_diagnostics,
+    _partition_with_target_k,
+    _physical_qubit_stats,
+    _probe_embedding,
+    _probe_embedding_with_estimate,
+    _summarize_partition_attempt,
+)
+from src.commands.poster_results.plotting import _plot_results
+from src.commands.poster_results.runner import (
+    FULL_TIMING_KEYS,
+    MR2S_TIMING_KEYS,
+    Mr2sOnlyPosterResultsRunner,
+    PosterResultsAggregator,
+    PosterResultsRunner,
+    PosterRunConfig,
+    RESULT_SERIES_KEYS,
+    TrialScheduler,
+    _merge_results_by_size,
+    _merge_nested_results_by_size,
+    _normalize_random_baseline,
+)
+from src.commands.poster_results.solvers import (
+    ALGORITHM_SOLVER_MAP,
+    ALL_ALGORITHMS,
+    BaseSolver,
+    DNC_STRATEGIES,
+    DncStrategySolver,
+    GlobalQuboSolver,
+    MR2S_VARIANTS,
+    Mr2sVariantSolver,
+    RandomBaselineSolver,
+    RawSASolver,
+    _as_finite_or_nan,
+    _build_edge_orienter,
+    _build_qubo_solver,
+    _build_sa_solver,
+    _extract_directed_edges_from_solution,
+    _flow_imbalance_score,
+    _mean_finite,
+    _run_mr2s_trial,
+    _run_poster_algorithm,
+    _run_trial,
+    _sample_random_orientations,
+    _trial_seed,
+)
+from src.commands.poster_results._cli import (
+    run,
+    run_mr2s_only,
+    _poster_trial_cache_key,
+    _poster_mr2s_trial_cache_key,
+    _calculate_random_baseline,
+    _solver_helpers,
+    _aggregate_mr2s_results,
+    register_parser,
+)
+# Override runner's _normalize_random_baseline with the CLI version that handles raw dicts.
+from src.commands.poster_results._cli import _normalize_random_baseline as _normalize_random_baseline
